% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{betas}
\alias{betas}
\title{Sample methylation data in betas format}
\format{
\subsection{\code{betas}}{

A data frame with 293 rows and 300 columns:
\describe{
\item{rows}{Illumina methylation array probes}
\item{columns}{TCGA-BRCA samples}
...
}
}
}
\source{
\url{https://portal.gdc.cancer.gov/projects/TCGA-BRCA}
}
\usage{
betas
}
\description{
A subset of methylation data from The Cancer Genome Atlas (TCGA)
breast cancer cohort collected using 450k methylation arrays
Report ...
}
\keyword{datasets}
