% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_regional_pcs.R
\name{compute_regional_pcs}
\alias{compute_regional_pcs}
\title{Compute regional PCs using methylation data for all regions included in the region map}
\usage{
compute_regional_pcs(meth, region_map, pc_method = "gd", verbose = FALSE)
}
\arguments{
\item{meth}{Data frame or matrix of methylation values to summarize; rows=CpGs, columns=samples}

\item{region_map}{Mapping of CpGs to regions, column 1 should be regions, column 2 should be CpGs with the same names as the rows of meth}

\item{pc_method}{String indicating the method for estimating dimension; "gd"=Gavish-Donoho (default), "mp"=Marchenko-Pastur}

\item{verbose}{Boolean; print output statements}
}
\value{
list containing PC results
}
\description{
Compute regional PCs using methylation data for all regions included in the region map
}
