% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_annots}
\alias{gene_annots}
\title{Annotations for gene promoter regions}
\format{
\subsection{\code{gene_annots}}{

A data frame with XXX rows and XXX columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
<>
}
\usage{
gene_annots
}
\description{
Annotations containing the start and end boundaries of
gene promoters for protein coding genes.
Report ...
}
\keyword{datasets}
