% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_regional_pcs.R
\name{summarize_region}
\alias{summarize_region}
\title{Summarize a region using regional principal components}
\usage{
summarize_region(region, region_map, meth, pc_method, verbose)
}
\arguments{
\item{region}{String; name of region being processed}

\item{region_map}{Data frame; Mapping of CpGs to regions, column 1 should be regions, column 2 should be CpGs with the same names as the rows of meth}

\item{meth}{Data frame or matrix; Methylation values to summarize; rows=CpGs, columns=samples}

\item{pc_method}{String; indicating the method for estimating dimension; "gd"=Gavish-Donoho (default), "mp"=Marchenko-Pastur}

\item{verbose}{Boolean; print output statements}
}
\value{
list containing PC results
}
\description{
Summarize a region using regional principal components
}
